posix.mak: Use the correct DMD when unittesting
There was a disconnect between posix.mak and test/common.mak on the BUILD value.
Because posix.mak defaults to BUILD=release, when unittest-debug was invoked,
the value of DMD would point to the release build of DMD, not debug.
This changes sets the BUILD value to its correct value so that the correct DMD is used.

Additionally, the 'target' was moved to 'unittest', as it is the main rule used:
when no BUILD is provided, it just recurses into calling itself twice,
once with BUILD=debug and once with BUILD=release.
When a build is provided (including after recursion), is actually does the test.
This was likely done as variables were "polluted" and a clean recursion makes
dealing with this much easier, hence why the dependency should only be done
after recursion, otherwise some of the variables are out of sync w.r.t. debug/release.