Move TraceInfo interface to core.internal.backtrace.primitives
This is a first step towards providing users the ability to inspect stack traces themselves.
The goal is to make this more user-friendly, by providing access to the underlying data,
such as address, function name, and file / line infos.

As this will require exposing the `Location` struct previously in `rt.backtrace.dwarf`,
and in order to avoid crowding `object` even more, this interface was moved to `primitives`,
where `Location` will be moved too.