Backtrace: Switch to dladdr to get the procedure name (libunwind)
This experimental libunwind feature (not enabled by default nor released yet)
has one big downside: it needs to allocate a very large buffer (128k) per trace handler.
This is because `libunwind` only allow to write to a buffer that is provided to it.
Since our symbols can be rather long, and we want to avoid GC allocations,
a static buffer was used. But even worse, when walking up the stack trace,
we have to eagerly store the function names, even if the user doesn't use this information.
This is a big waste of time and resources, and an issue that the backtrace implementation
doesn't share.
After investigation, it turns out that libunwind uses `dladdr` under the hood
(among other things, like string interning).
`dladdr` is available on all platforms we target, and provides us we a C string
on which we can use `strlen`, removing the need for those large buffers that are
eagerly filled.